requirements_new.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.6.0
sqlparse==0.5.3
tzdata==2025.2
Pillow==11.0.0

